requirements.txt
# requirements.txt
# ALL LIBRARIES ARE FREE AND OPEN SOURCE
# Install with: pip install -r requirements.txt

# ============================================================
# Data Collection (YouTube API)
# ============================================================
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# ============================================================
# Data Processing & Analysis
# ============================================================
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# ============================================================
# Machine Learning
# ============================================================
scikit-learn>=1.3.0
lightgbm>=4.0.0

# ============================================================
# Explainability
# ============================================================
shap>=0.44.0

# ============================================================
# Visualization
# ============================================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ============================================================
# Web App / Prototype
# ============================================================
streamlit>=1.28.0
plotly>=5.17.0>=1.28.0

# ============================================================
# Web Application Backend
# ============================================================
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# ============================================================
# Utilities
# ============================================================
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0
joblib>=1.3.0

# ============================================================
# Jupyter Notebooks
# ============================================================
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0
